+++ .storybook/preview.ts
@@ -1,6 +1,17 @@
 import type { Preview } from '@storybook/react';
+import { withThemeByClassName } from '@storybook/addon-themes';
+import '../src/styles/globals.css';
 
 const preview: Preview = {
+  decorators: [
+    withThemeByClassName({
+      themes: {
+        light: 'light',
+        dark: 'dark',
+      },
+      defaultTheme: 'light',
+    }),
+  ],
   parameters: {
     actions: { argTypesRegex: '^on[A-Z].*' },
     controls: {
@@ -9,6 +20,7 @@ const preview: Preview = {
         date: /Date$/,
       },
     },
+    layout: 'centered',
   },
 };